requirements.txt
beautifulsoup4==4.6.3
exifread==2.1.2
python-magic==0.4.15
requests==2.20.1
